# Add caching of flash reads