Add caching of flash reads from the device (#183), as the current metadata scans read from the same regions multiple times.

This provides a speedup to `picotool info -a` on my 4 partition device from 5.7s to 1.6s, and speeds up `picotool info` from 3.8s to 1.4s, so provides a good performance boost.

Gets fully invalidated on any writes, to minimise complexity, as the only read & write command is `picotool config` which wouldn't benefit from caching anyway. Also gets invalidated when using `picotool save --verify`, so you're not just verifying the cached data.